Tests cleanup
* Moved test_wait_for_stack_update to test_cluster_stack
* Moved test_update_stack_template to test_cluster_stack
* Moved test_stack_exists to test_aws_api
* Autoformat some test files.
