+++ cli/tests/awsbatch/test_awsbstat.py
@@ -4,8 +4,8 @@ import os
 import pytest
 
 from awsbatch import awsbstat
-from tests.utils import MockedBoto3Request, read_text
 from tests.conftest import DEFAULT_AWSBATCHCLICONFIG_MOCK_CONFIG
+from tests.utils import MockedBoto3Request, read_text
 
 ALL_JOB_STATUS = ["SUBMITTED", "PENDING", "RUNNABLE", "STARTING", "RUNNING", "SUCCEEDED", "FAILED"]
 DEFAULT_JOB_STATUS = ["SUBMITTED", "PENDING", "RUNNABLE", "STARTING", "RUNNING"]
