+++ cli/tests/pcluster/config/test_json_param_types.py
@@ -19,8 +19,8 @@ from pcluster.config.cfn_param_types import CfnSection
 from pcluster.config.mappings import CLUSTER_HIT
 from pcluster.config.param_types import SettingsParam, StorageData
 from pcluster.config.pcluster_config import PclusterConfig
-from tests.utils import MockedBoto3Request
 from tests.pcluster.config.utils import duplicate_config_file, get_mocked_pcluster_config
+from tests.utils import MockedBoto3Request
 
 # Mocked responses for describe_instance_types boto3 calls
 DESCRIBE_INSTANCE_TYPES_RESPONSES = {
