+++ cli/tests/pcluster/templates/test_imagebuilder_stack.py
@@ -14,8 +14,8 @@ from assertpy import assert_that
 
 import pcluster.utils as utils
 from pcluster.templates.cdk_builder import CDKTemplateBuilder
-
 from tests.common.dummy_aws_api import DummyAWSApi
+
 from ..models.imagebuilder_dummy_model import imagebuilder_factory
