+++ cli/tests/pcluster/test_utils.py
@@ -1,6 +1,5 @@
 """This module provides unit tests for the functions in the pcluster.utils module."""
 
-import json
 import logging
 import os
 from itertools import product
@@ -11,11 +10,8 @@ from assertpy import assert_that
 from botocore.exceptions import ClientError, EndpointConnectionError
 
 import pcluster.utils as utils
-from common.boto3.common import AWSClientError
-from pcluster.models.cluster import ClusterStack
 from pcluster.utils import Cache, get_bucket_url
 from tests.utils import MockedBoto3Request
-from tests.common.dummy_aws_api import DummyAWSApi
 
 FAKE_CLUSTER_NAME = "cluster-name"
 FAKE_STACK_NAME = utils.get_stack_name(FAKE_CLUSTER_NAME)
@@ -31,102 +27,6 @@ def test_get_stack_name():
     """Test utils.get_stack_name."""
     expected_stack_name = "parallelcluster-{0}".format(FAKE_CLUSTER_NAME)
     assert_that(utils.get_stack_name(FAKE_CLUSTER_NAME)).is_equal_to(expected_stack_name)
-
-
-@pytest.mark.parametrize(
-    "stack_statuses",
-    [
-        ["UPDATE_IN_PROGRESS", "UPDATE_IN_PROGRESS", "UPDATE_IN_PROGRESS", "UPDATE_COMPLETE"],
-        ["UPDATE_IN_PROGRESS", "UPDATE_IN_PROGRESS", "UPDATE_IN_PROGRESS", "anything other than UPDATE_IN_PROGRESS"],
-        ["UPDATE_COMPLETE"],
-    ],
-)
-def test_wait_for_stack_update(mocker, stack_statuses):
-    """
-    Verify that ClusterStack._wait_for_update behaves as expected.
-
-    _wait_for_update should call updated_status until the StackStatus is anything besides UPDATE_IN_PROGRESS,
-    use that to get expected call count for updated_status
-    """
-    expected_call_count = 0
-    for status_idx, status in enumerate(stack_statuses):
-        if status != "UPDATE_IN_PROGRESS":
-            expected_call_count = status_idx + 1
-            break
-
-    cluster_stack = ClusterStack({"StackName": FAKE_STACK_NAME})
-
-    updated_status_mock = mocker.patch.object(cluster_stack, "updated_status", side_effect=stack_statuses)
-    mocker.patch("pcluster.models.cluster.time.sleep")  # so we don't actually have to wait
-
-    cluster_stack._wait_for_update()
-    assert_that(updated_status_mock.call_count).is_equal_to(expected_call_count)
-
-
-@pytest.mark.parametrize(
-    "error_message",
-    [
-        None,
-        "No UpDatES ARE TO BE PERformed",
-        "some longer message also containing no updates are to be performed and more words at the end"
-        "some other error message",
-    ],
-)
-def test_update_stack_template(mocker, error_message):
-    """Verify that utils.update_stack_template behaves as expected."""
-    template_body = {"TemplateKey": "TemplateValue"}
-    cfn_params = [{"ParameterKey": "Key", "ParameterValue": "Value"}]
-    response = error_message or {"StackId": "stack ID"}
-
-    mocker.patch("common.aws.aws_api.AWSApi.instance", return_value=DummyAWSApi())
-    mocker.patch("common.boto3.cfn.CfnClient.get_stack_template", return_value=template_body)
-    mocker.patch(
-        "common.boto3.cfn.CfnClient.update_stack",
-        return_value=response,
-        expected_params={
-            "stack_name": FAKE_STACK_NAME,
-            "updated_template": json.dumps(template_body, indent=2),
-            "params": cfn_params,
-        },
-        side_effect=AWSClientError(function_name="update_stack", message=error_message)
-        if error_message is not None
-        else None,
-    )
-
-    cluster_stack = ClusterStack({"StackName": FAKE_STACK_NAME})
-    wait_for_update_mock = mocker.patch.object(cluster_stack, "_wait_for_update")
-
-    if error_message is None or "no updates are to be performed" in error_message.lower():
-        cluster_stack._update_template()
-        if error_message is None or "no updates are to be performed" not in error_message.lower():
-            assert_that(wait_for_update_mock.called).is_true()
-        else:
-            assert_that(wait_for_update_mock.called).is_false()
-    else:
-        full_error_message = "Unable to update stack template for stack {stack_name}: {emsg}".format(
-            stack_name=FAKE_STACK_NAME, emsg=error_message
-        )
-        with pytest.raises(AWSClientError, match=full_error_message) as sysexit:
-            cluster_stack._update_template()
-        assert_that(sysexit.value.code).is_not_equal_to(0)
-
-
-@pytest.mark.parametrize(
-    "response,is_error",
-    [
-        (
-            AWSClientError(function_name="describe_stack", message=f"Stack with id {FAKE_STACK_NAME} does not exist"),
-            True,
-        ),
-        ({"Stacks": [{"StackName": FAKE_STACK_NAME, "CreationTime": 0, "StackStatus": "CREATED"}]}, False),
-    ],
-)
-def test_stack_exists(mocker, response, is_error):
-    """Verify that CfnClient.stack_exists behaves as expected."""
-    should_exist = not is_error
-    mocker.patch("common.aws.aws_api.AWSApi.instance", return_value=DummyAWSApi())
-    mocker.patch("common.boto3.cfn.CfnClient.describe_stack", side_effect=response)
-    assert_that(DummyAWSApi().instance().cfn.stack_exists(FAKE_STACK_NAME)).is_equal_to(should_exist)
 
 
 @pytest.mark.parametrize(
