+++ cli/tests/pcluster/validators/test_cluster_validators.py
@@ -34,8 +34,8 @@ from pcluster.validators.cluster_validators import (
     SchedulerOsValidator,
     TagKeyValidator,
 )
-from tests.utils import MockedBoto3Request
 from tests.pcluster.validators.utils import assert_failure_messages
+from tests.utils import MockedBoto3Request
 
 
 @pytest.fixture()
