+++ cli/tests/pcluster/validators/test_ebs_validators.py
@@ -22,8 +22,8 @@ from pcluster.validators.ebs_validators import (
     EbsVolumeTypeSizeValidator,
     SharedEBSVolumeIdValidator,
 )
-from tests.utils import MockedBoto3Request
 from tests.pcluster.validators.utils import assert_failure_messages
+from tests.utils import MockedBoto3Request
 
 
 @pytest.fixture()
