+++ cli/tests/pcluster/validators/test_fsx_validators.py
@@ -21,8 +21,8 @@ from pcluster.validators.fsx_validators import (
     FsxStorageCapacityValidator,
     FsxStorageTypeOptionsValidator,
 )
-from tests.utils import MockedBoto3Request
 from tests.pcluster.validators.utils import assert_failure_messages
+from tests.utils import MockedBoto3Request
 
 
 @pytest.fixture()
